% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed2.R
\name{speed2run}
\alias{speed2run}
\title{Test pathway gene signature enrichment using SPEED2}
\usage{
speed2run(genes, background_genes = c(), shuffle = F, custom_signatures = NULL)
}
\arguments{
\item{genes}{list of genes in Entrez or gene symbols format to test
for enrichment in SPEED2. Maximum 500 genes.}

\item{background_genes}{(optional) list of background genes in Entrez or gene
symbol from which \code{genes} were selected. If not provided, the full
set of SPEED2 genes are used.}

\item{shuffle}{(optional) shuffle identities of genes in SPEED2, for control
experiments.}

\item{custom_signatures}{(optional, advanced) user provided custom pathway
gene signatures to use instead of SPEED2, provided as a tibble with the
following columns: p_id (pathway id), g_id (gene id), zrank_signed_mean
(average normalized z score across many experiments, between -1 and +1),
P.bates (p-value associated with zrank_signed_mean). Users are expected
to handle Entrez conversion and background gene list themselves.}
}
\value{
List with four items: \code{df_stattest} a tibble with enrichment
    scores, \code{df_rankcoords} coordinates for GSEA plot (see publication),
    and two lists with unmatched items in \code{genes} and
    \code{background_genes}.
}
\description{
Test pathway gene signature enrichment using SPEED2
}
\examples{
ret = speed2run(genes=speed2:::speed2_signatures$g_id[1:50])
}
